[DRAFT] Move parameter sweep to general case
